+++ packages/utils/src/lib/reports/constants.ts
@@ -12,7 +12,7 @@ export const SCORE_COLOR_RANGE = {
 /* eslint-enable no-magic-numbers */
 
 export const CATEGORIES_TITLE = '🏷 Categories';
-export const FOOTER_PREFIX = 'Made with ❤ by'; // replace ❤️ with ❤, because of ❤️ has output issues
+export const FOOTER_PREFIX = 'Made with ❤ by'; // replace ❤️ with ❤, because ❤️ has output issues
 export const CODE_PUSHUP_DOMAIN = 'code-pushup.dev';
 export const README_LINK = 'https://github.com/code-pushup/cli#readme';
 export const reportHeadlineText = 'Code PushUp Report';
