+++ packages/utils/src/lib/reports/formatting.ts
@@ -1,5 +1,14 @@
-import { Table } from '@code-pushup/models';
-import { Hierarchy, headline, paragraphs, tableMd } from './md';
+import { AuditReport, Table } from '@code-pushup/models';
+import {
+  Hierarchy,
+  NEW_LINE,
+  SPACE,
+  headline,
+  link,
+  paragraphs,
+  section,
+  tableMd,
+} from './md';
 
 export function tableSection(
   table: Table | undefined,
@@ -24,4 +33,25 @@ export function tableSection(
   const render = (h: string, l: Hierarchy | 0) =>
     l === 0 ? heading : headline(h, l);
   return paragraphs(heading ? render(heading, level) : false, tableMd(table));
+}
+
+// @TODO extract `Pick<AuditReport, 'docsUrl' | 'description'>` to a reusable schema and type
+export function metaDescription({
+  docsUrl,
+  description,
+}: Pick<AuditReport, 'docsUrl' | 'description'>): string {
+  if (docsUrl) {
+    const docsLink = link(docsUrl, '📖 Docs');
+    if (!description) {
+      return section(docsLink);
+    }
+    const parsedDescription = description.toString().endsWith('```')
+      ? `${description}${NEW_LINE + NEW_LINE}`
+      : `${description}${SPACE}`;
+    return section(`${parsedDescription}${docsLink}`);
+  }
+  if (description && description.trim().length > 0) {
+    return section(description);
+  }
+  return '';
 }
