+++ packages/utils/src/lib/reports/formatting.unit.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it } from 'vitest';
-import { tableSection } from './formatting';
+import { metaDescription, tableSection } from './formatting';
+import { NEW_LINE } from './md';
 
 describe('tableSection', () => {
   it('should render complete section', () => {
@@ -38,5 +39,49 @@ describe('tableSection', () => {
         { heading: 'LCP Breakdown', level: 3 },
       ),
     ).toMatchSnapshot();
+  });
+});
+
+describe('metaDescription', () => {
+  it('should return empty string if no options are given', () => {
+    expect(metaDescription({})).toBe('');
+  });
+
+  it('should return description if only description is given', () => {
+    expect(
+      metaDescription({
+        description: 'Audit to track bundle size',
+      }),
+    ).toBe(`Audit to track bundle size${NEW_LINE}`);
+  });
+
+  it('should return docsUrl if only docsUrl is given', () => {
+    expect(
+      metaDescription({
+        docsUrl: `http://code-pushup.dev/audits/#lcp`,
+      }),
+    ).toBe(`[📖 Docs](http://code-pushup.dev/audits/#lcp)${NEW_LINE}`);
+  });
+
+  it('should docs and description if both given', () => {
+    expect(
+      metaDescription({
+        description: 'Audit for loading performance',
+        docsUrl: 'http://code-pushup.dev/audits/#lcp',
+      }),
+    ).toBe(
+      `Audit for loading performance [📖 Docs](http://code-pushup.dev/audits/#lcp)${NEW_LINE}`,
+    );
+  });
+
+  it('should have a NEW_LINE if description ends with a code block', () => {
+    expect(
+      metaDescription({
+        description: 'Audit to loading performance```',
+        docsUrl: 'http://code-pushup.dev/audits/#lcp',
+      }),
+    ).toBe(
+      `Audit to loading performance\`\`\`${NEW_LINE}${NEW_LINE}[📖 Docs](http://code-pushup.dev/audits/#lcp)${NEW_LINE}`,
+    );
   });
 });
