+++ packages/utils/src/lib/reports/generate-md-report.ts
@@ -1,7 +1,6 @@
 import { AuditReport, Issue, Report, Table } from '@code-pushup/models';
-import { formatDate, formatDuration, slugify } from '../formatting';
+import { formatDate, formatDuration } from '../formatting';
 import {
-  CATEGORIES_TITLE,
   FOOTER_PREFIX,
   README_LINK,
   issuesTableHeadings,
@@ -10,144 +9,31 @@ import {
   reportHeadlineText,
   reportMetaTableAlignment,
   reportMetaTableHeaders,
-  reportOverviewTableAlignment,
-  reportOverviewTableHeaders,
 } from './constants';
-import { tableSection } from './formatting';
+import { metaDescription, tableSection } from './formatting';
+import {
+  categoriesDetailsSection,
+  categoriesOverviewSection,
+} from './generate-md-report-categoy-section';
 import { style as htmlFontStyle } from './html/font-style';
 import {
-  NEW_LINE,
   SPACE,
   details,
   h1,
   h2,
   h3,
-  indentation,
-  li,
   link,
   paragraphs,
   section,
   style,
 } from './md';
-import { ScoredGroup, ScoredReport } from './types';
+import { ScoredReport } from './types';
 import {
-  countCategoryAudits,
   formatReportScore,
   getPluginNameFromSlug,
-  getSortableAuditByRef,
-  getSortableGroupByRef,
   scoreMarker,
   severityMarker,
 } from './utils';
-
-export function reportOverviewSection(
-  report: Pick<ScoredReport, 'categories' | 'plugins'>,
-): string {
-  const { categories, plugins } = report;
-  if (categories.length > 0 && plugins.length > 0) {
-    const tableContent: Table = {
-      headings: reportOverviewTableHeaders,
-      rows: categories.map(({ title, refs, score }) => ({
-        // @TODO shouldn't this be the category slug as title is not unique? => slugify(title) -> slug
-        category: link(`#${slugify(title)}`, title),
-        score: `${scoreMarker(score)}${SPACE}${style(
-          formatReportScore(score),
-        )}`,
-        audits: countCategoryAudits(refs, plugins).toString(),
-      })),
-      alignment: reportOverviewTableAlignment,
-    };
-    return tableSection(tableContent);
-  }
-  return '';
-}
-
-export function categoriesDetailsSection(
-  report: Pick<ScoredReport, 'categories' | 'plugins'>,
-): string {
-  const { categories, plugins } = report;
-
-  const categoryDetails = categories.flatMap(category => {
-    const categoryTitle = h3(category.title);
-    const categoryScore = `${scoreMarker(
-      category.score,
-    )}${SPACE}Score:  ${style(formatReportScore(category.score))}`;
-
-    const categoryMDItems = category.refs.map(ref => {
-      // Add group details
-      if (ref.type === 'group') {
-        const group = getSortableGroupByRef(ref, plugins);
-        const groupAudits = group.refs.map(groupRef =>
-          getSortableAuditByRef(
-            { ...groupRef, plugin: group.plugin, type: 'audit' },
-            plugins,
-          ),
-        );
-        const pluginTitle = getPluginNameFromSlug(ref.plugin, plugins);
-        return paragraphs(categoryGroupItem(group, groupAudits, pluginTitle));
-      }
-      // Add audit details
-      else {
-        const audit = getSortableAuditByRef(ref, plugins);
-        const pluginTitle = getPluginNameFromSlug(ref.plugin, plugins);
-        return paragraphs(categoryRef(audit, pluginTitle));
-      }
-    });
-
-    return section(
-      categoryTitle,
-      metaDescription(category),
-      categoryScore,
-      ...categoryMDItems,
-    );
-  });
-
-  return paragraphs(h2(CATEGORIES_TITLE), ...categoryDetails);
-}
-
-export function categoryRef(
-  { title, score, value, displayValue }: AuditReport,
-  pluginTitle: string,
-): string {
-  const auditTitleAsLink = link(
-    `#${slugify(title)}-${slugify(pluginTitle)}`,
-    title,
-  );
-  const marker = scoreMarker(score, 'square');
-  return li(
-    `${marker}${SPACE}${auditTitleAsLink}${SPACE}(_${pluginTitle}_) - ${style(
-      (displayValue || value).toString(),
-    )}`,
-  );
-}
-
-export function categoryGroupItem(
-  { score = 0, title }: ScoredGroup,
-  groupAudits: AuditReport[],
-  pluginTitle: string,
-): string {
-  const groupTitle = li(
-    `${scoreMarker(score)}${SPACE}${title}${SPACE}(_${pluginTitle}_)`,
-  );
-  const auditTitles = groupAudits.map(
-    ({ title: auditTitle, score: auditScore, value, displayValue }) => {
-      const auditTitleLink = link(
-        `#${slugify(auditTitle)}-${slugify(pluginTitle)}`,
-        auditTitle,
-      );
-      const marker = scoreMarker(auditScore, 'square');
-      return indentation(
-        li(
-          `${marker}${SPACE}${auditTitleLink} - ${style(
-            String(displayValue ?? value),
-          )}`,
-        ),
-      );
-    },
-  );
-
-  return paragraphs(groupTitle, ...auditTitles);
-}
 
 export function auditDetailsAuditValue({
   score,
@@ -164,7 +50,7 @@ export function generateMdReport(report: ScoredReport): string {
 
   return paragraphs(
     h1(reportHeadlineText),
-    printCategories ? reportOverviewSection(report) : '',
+    printCategories ? categoriesOverviewSection(report) : '',
     printCategories ? categoriesDetailsSection(report) : '',
     auditsSection(report),
     aboutSection(report),
@@ -224,27 +110,6 @@ export function auditDetails(audit: AuditReport) {
   );
 }
 
-// @TODO extract `Pick<AuditReport, 'docsUrl' | 'description'>` to a reusable schema and type
-export function metaDescription({
-  docsUrl,
-  description,
-}: Pick<AuditReport, 'docsUrl' | 'description'>): string {
-  if (docsUrl) {
-    const docsLink = link(docsUrl, '📖 Docs');
-    if (!description) {
-      return section(docsLink);
-    }
-    const parsedDescription = description.toString().endsWith('```')
-      ? `${description}${NEW_LINE + NEW_LINE}`
-      : `${description}${SPACE}`;
-    return section(`${parsedDescription}${docsLink}`);
-  }
-  if (description && description.trim().length > 0) {
-    return section(description);
-  }
-  return '';
-}
-
 export function auditsSection({
   plugins,
 }: Pick<ScoredReport, 'plugins'>): string {
@@ -267,16 +132,15 @@ export function aboutSection(
   report: Omit<ScoredReport, 'packageName'>,
 ): string {
   const { date, plugins } = report;
-
-  const formattedDate = formatDate(new Date(date));
   const reportMetaTable = reportMetaData(report);
-
   const pluginMetaTable = reportPluginMeta({ plugins });
   const level = 3;
   return paragraphs(
     h2('About'),
     section(
-      `Report was created by [Code PushUp](${README_LINK}) on ${formattedDate}.`,
+      `Report was created by [Code PushUp](${README_LINK}) on ${formatDate(
+        new Date(date),
+      )}.`,
     ),
     tableSection(reportMetaTable, { heading: 'Report overview:', level }),
     tableSection(pluginMetaTable, { heading: 'Plugins overview:', level }),
