+++ packages/utils/src/lib/reports/generate-md-report.unit.test.ts
@@ -7,15 +7,10 @@ import {
   auditDetailsAuditValue,
   auditDetailsIssues,
   auditsSection,
-  categoriesDetailsSection,
-  categoryGroupItem,
-  categoryRef,
   generateMdReport,
-  metaDescription,
-  reportOverviewSection,
 } from './generate-md-report';
-import { NEW_LINE } from './md/constants';
-import { ScoredGroup, ScoredReport } from './types';
+import { NEW_LINE } from './md';
+import { ScoredReport } from './types';
 
 const baseScoredReport = {
   date: '2025.01.01',
@@ -52,235 +47,6 @@ const baseScoredReport = {
     },
   ],
 } as ScoredReport;
-
-describe('metaDescription', () => {
-  it('should return empty string if no options are given', () => {
-    expect(metaDescription({})).toBe('');
-  });
-
-  it('should return description if only description is given', () => {
-    expect(
-      metaDescription({
-        description: 'Audit to track bundle size',
-      }),
-    ).toBe(`Audit to track bundle size${NEW_LINE}`);
-  });
-
-  it('should return docsUrl if only docsUrl is given', () => {
-    expect(
-      metaDescription({
-        docsUrl: `http://code-pushup.dev/audits/#lcp`,
-      }),
-    ).toBe(`[📖 Docs](http://code-pushup.dev/audits/#lcp)${NEW_LINE}`);
-  });
-
-  it('should docs and description if both given', () => {
-    expect(
-      metaDescription({
-        description: 'Audit for loading performance',
-        docsUrl: 'http://code-pushup.dev/audits/#lcp',
-      }),
-    ).toBe(
-      `Audit for loading performance [📖 Docs](http://code-pushup.dev/audits/#lcp)${NEW_LINE}`,
-    );
-  });
-
-  it('should have a NEW_LINE if description ends with a code block', () => {
-    expect(
-      metaDescription({
-        description: 'Audit to loading performance```',
-        docsUrl: 'http://code-pushup.dev/audits/#lcp',
-      }),
-    ).toBe(
-      `Audit to loading performance\`\`\`${NEW_LINE}${NEW_LINE}[📖 Docs](http://code-pushup.dev/audits/#lcp)${NEW_LINE}`,
-    );
-  });
-});
-
-// === Categories Overview Section
-
-describe('categoriesSection', () => {
-  it('should NOT render categories table', () => {
-    const md = reportOverviewSection({
-      plugins: [],
-      categories: [],
-    } as unknown as ScoredReport);
-    expect(md).toBe('');
-  });
-
-  it('should render complete categories table', () => {
-    expect(
-      reportOverviewSection({
-        plugins: [
-          {
-            slug: 'eslint',
-            title: 'Eslint',
-          },
-          {
-            slug: 'lighthouse',
-            title: 'Lighthouse',
-          },
-        ],
-        categories: [
-          {
-            slug: 'bug-prevention',
-            title: 'Bug Prevention',
-            score: 0.98,
-            refs: [{ slug: 'no-let', type: 'audit' }],
-          },
-          {
-            slug: 'performance',
-            title: 'Performance',
-            score: 0.74,
-            refs: [{ slug: 'largest-contentful-paint', type: 'audit' }],
-          },
-          {
-            slug: 'typescript',
-            title: 'Typescript',
-            score: 0.14,
-            refs: [{ slug: 'no-any', type: 'audit' }],
-          },
-        ],
-      } as unknown as ScoredReport),
-    ).toMatchSnapshot();
-  });
-});
-
-// === Categories Details
-
-describe('categoryRef', () => {
-  it('should render partial category reference', () => {
-    expect(
-      categoryRef(
-        {
-          slug: 'score-report-performance',
-          title: 'Score Report Performance',
-          value: 12_245,
-          score: 1,
-        },
-        'lighthouse',
-      ),
-    ).toBe(
-      '- 🟩 [Score Report Performance](#score-report-performance-lighthouse) (_lighthouse_) - **12245**',
-    );
-  });
-
-  it('should render complete category reference', () => {
-    expect(
-      categoryRef(
-        {
-          slug: 'score-report-performance',
-          title: 'Score Report Performance',
-          value: 1,
-          score: 0,
-          displayValue: '12 errors',
-        },
-        'lighthouse',
-      ),
-    ).toBe(
-      '- 🟥 [Score Report Performance](#score-report-performance-lighthouse) (_lighthouse_) - **12 errors**',
-    );
-  });
-});
-
-describe('categoryGroupItem', () => {
-  it('should render partial category reference', () => {
-    expect(
-      categoryGroupItem(
-        {
-          slug: 'bug-prevention',
-          title: 'Bug Prevention',
-          score: 0.9,
-        } as ScoredGroup,
-        [
-          { title: 'No let', slug: 'no-let', score: 0, value: 23 },
-          { title: 'No any', slug: 'no-any', score: 0.6, value: 91 },
-        ],
-        'Eslint',
-      ),
-    ).toMatchSnapshot('');
-  });
-
-  it('should render complete category reference', () => {
-    expect(
-      categoryGroupItem(
-        {
-          slug: 'bug-prevention',
-          title: 'Bug Prevention',
-          score: 0.6,
-        } as ScoredGroup,
-        [
-          {
-            title: 'No any',
-            slug: 'no-any',
-            score: 0,
-            value: 12,
-            displayValue: '12 errors',
-          },
-          { title: 'No let', slug: 'no-let', score: 1, value: 0 },
-        ],
-        'Eslint',
-      ),
-    ).toMatchSnapshot();
-  });
-});
-
-describe('categoriesDetails', () => {
-  it('should render complete categories details', () => {
-    expect(
-      categoriesDetailsSection({
-        plugins: [
-          {
-            slug: 'eslint',
-            title: 'Eslint',
-            audits: [
-              { slug: 'no-let', title: 'No let', score: 0, value: 5 },
-              { slug: 'no-any', title: 'No any', score: 1, value: 0 },
-            ],
-          },
-          {
-            slug: 'lighthouse',
-            title: 'Lighthouse',
-            audits: [
-              {
-                slug: 'largest-contentful-paint',
-                title: 'Largest Contentful Paint',
-                score: 0.7,
-                value: 2905,
-              },
-            ],
-          },
-        ],
-        categories: [
-          {
-            slug: 'bug-prevention',
-            title: 'Bug Prevention',
-            score: 0.98,
-            refs: [{ slug: 'no-let', type: 'audit', plugin: 'eslint' }],
-          },
-          {
-            slug: 'performance',
-            title: 'Performance',
-            score: 0.74,
-            refs: [
-              {
-                slug: 'largest-contentful-paint',
-                type: 'audit',
-                plugin: 'lighthouse',
-              },
-            ],
-          },
-          {
-            slug: 'typescript',
-            title: 'Typescript',
-            score: 0.14,
-            refs: [{ slug: 'no-any', type: 'audit', plugin: 'eslint' }],
-          },
-        ],
-      } as unknown as ScoredReport),
-    ).toMatchSnapshot();
-  });
-});
 
 // === Audit Details
